Fix race condition on lazy constructor resolution
To resolve cached NSString and NSNumber constructors, the cached class was checked for being `null`,
not the constructor itself. This made possible for two racing threads to cause a `call` on a null function,
if the first gets preempted before assigning the value to the constructor variable, but after setting the
cached class variable.

Checking instead for the constructor variable avoids the possibility of calling a null function.
The performance cost of this is that both the competing threads will resolve the constructor,
but the impact is limited in time (until the resolution happens once).