Fix CModule lifetime issue
Merely being available to the closure below it is not enough if the
closure never uses it. This happened to work on Duktape due to its
simplistic garbage collector, but was an issue on V8 as it's able to
determine that “cm” is unused.

Kudos to @hexploitable for helping track this one down.