Fix ObjC.choose() on latest TinyCC
By supplying missing symbols when constructing the CModule used for
fastpaths.

This bug was hidden by the previous version of TinyCC, which used
dlsym() as a fallback on UNIX platforms. The latest TinyCC no longer
does this when using “-nostdlib”, which we are using. The previous
behavior was however an unintended feature that made it hard to write
portable scripts, so we will not bring it back.