Handle extended block type encoding
This change adds basic support for extended block type encoding. When
that is enabled by the compiler, block type encodings also include the
types for the arguments, enclosed in `<>` immediately following the
block identifier `@?`.

Here "basic" means that now `readType()` just skips the additional
type annotations of block arguments (which could be nested) instead of
throwing `"Unable to handle type <"`.